+++ Services/UserService.cs
@@ -23,7 +23,9 @@ namespace PhotoScavengerHunt.Services
             try
             {
                 await _userRepo.EnsureUsernameIsValidAsync(name);
-                await _userRepo.EnsureAgeIsValidAsync(age);
+                
+                if (age <= 0 || age > 125)
+                    throw new ArgumentException("Invalid age value.");
 
                 var profile = new UserProfile
                 {
@@ -70,7 +72,8 @@ namespace PhotoScavengerHunt.Services
         {
             try
             {
-                await _userRepo.EnsureUserExistsAsync(id);
+                if(!await _userRepo.ExistsAsync(id))
+                throw new ChallengeNotFoundException("User does not exist.");
                 var user = await _userRepo.GetByIdAsync(id);
                 return (true, string.Empty, user);
             }
@@ -89,7 +92,9 @@ namespace PhotoScavengerHunt.Services
         {
             try
             {
-                var user = await _userRepo.GetByIdOrThrowAsync(id);
+                var user = await _userRepo.GetByIdAsync(id);
+                if (user == null)
+                    throw new ChallengeNotFoundException("User not found.");
                 await _userRepo.RemoveAsync(user);
                 await _userRepo.SaveChangesAsync();
                 return (true, string.Empty);
